requirements.txt
Django==2.2.27
django-environ==0.4.4
social-auth-app-django==2.1.0
